Refactor; split Showcase RSS module into seperate file.
